+++ drake/systems/framework/primitives/test/CMakeLists.txt
@@ -1,14 +1,8 @@
 drake_add_cc_test(adder_test)
 target_link_libraries(adder_test drakeSystemFramework)
 
-drake_add_cc_test(pass_through_test)
-target_link_libraries(pass_through_test drakeSystemFramework)
-
-drake_add_cc_test(pass_through_scalartype_test)
-target_link_libraries(pass_through_scalartype_test drakeSystemFramework)
-
-drake_add_cc_test(integrator_test)
-target_link_libraries(integrator_test drakeSystemFramework)
+drake_add_cc_test(affine_system_test)
+target_link_libraries(affine_system_test drakeSystemFramework)
 
 drake_add_cc_test(constant_value_source_test)
 target_link_libraries(constant_value_source_test drakeSystemFramework)
@@ -16,20 +10,35 @@ target_link_libraries(constant_value_source_test drakeSystemFramework)
 drake_add_cc_test(constant_vector_source_test)
 target_link_libraries(constant_vector_source_test drakeSystemFramework)
 
+drake_add_cc_test(demultiplexer_test)
+target_link_libraries(demultiplexer_test drakeSystemFramework)
+
 drake_add_cc_test(gain_test)
 target_link_libraries(gain_test drakeSystemFramework)
 
 drake_add_cc_test(gain_scalartype_test)
 target_link_libraries(gain_scalartype_test drakeSystemFramework)
 
-drake_add_cc_test(demultiplexer_test)
-target_link_libraries(demultiplexer_test drakeSystemFramework)
+drake_add_cc_test(linear_system_test)
+target_link_libraries(linear_system_test drakeSystemFramework)
+
+drake_add_cc_test(integrator_test)
+target_link_libraries(integrator_test drakeSystemFramework)
 
 drake_add_cc_test(multiplexer_test)
 target_link_libraries(multiplexer_test drakeSystemFramework)
 
 drake_add_cc_test(trajectory_source_test)
 target_link_libraries(trajectory_source_test drakeSystemFramework)
+
+drake_add_cc_test(pass_through_test)
+target_link_libraries(pass_through_test drakeSystemFramework)
+
+drake_add_cc_test(pass_through_scalartype_test)
+target_link_libraries(pass_through_scalartype_test drakeSystemFramework)
+
+drake_add_cc_test(time_varying_polynomial_source_test)
+target_link_libraries(time_varying_polynomial_source_test drakeSystemFramework)
 
 drake_add_cc_test(zero_order_hold_test)
 target_link_libraries(zero_order_hold_test drakeSystemFramework)
